CORE/webscraping/requirements.txt
psycopg2==2.7.1
newspaper3k
celery
tenacity
sqlalchemy
